[shravanthi] - Load csv data, Calculated mean of age and salary

+++ DataPreProcessingOne/PurchaseList.cs
@@ -8,10 +8,10 @@ namespace DataPreProcessingOne
         public string Country { get; set; }
 
         [LoadColumn(1)]
-        public int Age { get; set; }
+        public float Age { get; set; }
 
         [LoadColumn(2)]
-        public int Salary { get; set; }
+        public float Salary { get; set; }
 
         [LoadColumn(3)]
         public string Purchased { get; set; }
